requirements.txt
pandas
numpy
scikit-learn
scipy
streamlit
bcrypt
tensorflow
matplotlib
python-dotenv